requirements.txt
PyQt6
pyqt-tools
PyMySQL
mysql-connector-python